Continued after last commit, fixed bugs introduced
(This message is a test.) I am in the process of making a heierarchical page number
system that will allow me to point users to such things as biblical chapter and verse
and drama act and scene. This is a complicated process and this commit continues on the
work performed in the last commit toward this end.