Created ahash() for annotation indicators
But I still don't like it. The issue seems to be because, for any given
line, if there are more than one annotations, they are inserted into the
line in a descending order, starting from the end of the line working
towards the beginning. This prevents the destruction of the line's char
index so that when I enter later annotations in the line, they aren't
inserted too early because the index shifted when I entered earlier
annotations (because we're inserting a bracket annotation within <sup>
and <a href> tags we end up seriously increasing the length of the
actual line with each insertion). I cannot seem to solve this problem.
Note, however, that for some reason I haven't seen a change in indexing
when I change the order_by clause on the last_char_idx between asc() and
desc(), which makes absolutely not a lick of sense at all.