Converted Line model to include em_status
Also had to create insertlines.py as a python script to use sqlalchemy
to insert the lines into mysql. It has a dumb feature that it queries
the database twice for the id of kind with every line. I can easily make
one query, index them all into a dictionary, and that would be that.
I'll do that.