I don't want migrations being committed anymore
They frequently have issues so I don't want them a part of my
repository. They're autogenerated based on code that already exists and
I frequently delete them and regenerate them to avoid various problems.
I'm unsure if I want to eliminate migrations or not, I do not know much
about alembic and so am going to stick with the current system for now.