Created book summary edit.
There was a rather strange bug on this because of the indexing. Every
time we went to modify the summary SQLAlchemy was throwing an error
because we were issuing SQL to access a certain attribute before
committing changes. Had to modify Book.author to be a joined eager load.
That solved the problem