Created a populate and repoopulate script to automate repopulating the database
Holy crap was that getting annoying, and then having to recreate the
same users all the time. this does it for me, but I'm not adding it to
the system; it requires me using the actual path of the files; this is
not ideal because it's not portable. That said, the shebangs in my
insert scripts do the same thing; those will have to be adjusted down
the road.