Improved notifications
It's really dumb to communicate reputation changes on notifications, as
these don't apply to everyone. Now the notifications only work for edit
approval and rejection (and the notifiers differ in each case).

I will create a reputation history view for users that queries their own
ReputationChanges. That makes a ton more sense.

Notifications have to be created for a lot of follower stuff.