Improve writers index and writer view page
Had to add new relationships to writer (namely, `authored`, `edited`, and
`translated`) so that we can keep those categories separate.

I added sort categories corresponding to these things in the index and I
added those categories to the writer page. I also made the authorship on
the actual book page scroll through all the authors that we can have.

On top of all that, I refined the styles slightly corresponding to
cards. That said, the whole site needs to be retooled anyway, so that
was kind of a waste of time