Add `edition_id` to `Edit`
It solves so many problems, it makes a lot of things way easier, and
there is virtually no reason (save, maybe, excessive database
normalization) to eliminate it.

I got sick and tired of having to join `Annotation` every time I needed
to query for annotations within an edition or text. I should be able to
fix some routes thanks to this.

It also speeds up `context` and `selection` generation like crazy