Fix for "commited state" error in `SearchableMixin`
This should be it. Like the end of that bug. Essentially, I just run
through the session changes before the commit, running `getattr()` on
all the necessary fields so that they get loaded into memory before the
commit. That way I don't have to have them loaded `lazy="joined"` (which
gets to be a headache).