Eliminated rollback method
I don't believe in it any more. There is no point. The edits are
approved by a panel of high rep users; therefore there shouldn't be any
reason to roll it back. That makes complete sense in Wikipedia, where
edits are _not_ approved, but no sense here. This is precisely why there
is no rollback history on StackOverflow