Improve user authorization system
N.B. I may very well rename this system. I don't know that it's actually
administration because users will be able to review edits.

I have removed the config AUTHORZATION dictionary that defined
reputations in favor of a column on `Right` that defines min_rep. If
there is none, it defaults to NULL and therefore will require the token
to be in the user's rights.

I have created a wrapper function for authorizing routes, so I don't
call the authorize method all the time. It must come after a
login_required, or else there would be no point anyway.

I am of two minds about certain reputation authorizations. On the one
hand, I am thinking that I want edits and annotations to require a
minimum reputation and encourage discussion as the initial route to
increase rep. I am more inclined to do this for creating annotations
than edits. On the other, I want anonymous users to be able to make
edits. I am slowly leaning toward the former.