Add texts to this repository
This is redundant, as I maintain copies in a separate repo. That will
continue for the foreseeable future. But for right now I want these in
this repo for various reasons including but not limited to the
self-contained nature of the app using the scripts.

Perhaps eventually I will use these files solely for testing.

That will probably be what I do.