Convert processor to output json
I actually didn't have much trouble on this edit. First I made `lout()`
return a dictionary instead of write to `fout` a @-separated string.
Then I made a wrapper function that took the result from the function
and added it to a global list called `data_array`. Then I wrapped
`lout()` and at the end of the file called
`fout.write(json.dumps(data_array))` and there we have it.

Must do this for `preprocessor.py` and convert `insertlines.py` and
`insertannotations.py` to take the format.