Convert all shebangs to pull from environment variable
I love unix.stackexchange.com. It is incredible.

Gilles posted this in 2011!

>Here's one way to achieve the same effect in Python. In the shell,
>"true" is the true utility, which ignores its arguments (two
>single-character strings : and ') and returns a true value. In Python,
>"true" is a string which is true when interpreted as a boolean, so this
>is an if instruction that's always true and executes a string literal.

    #!/bin/sh
    if "true" : '''\'
    then
    exec "$FOO/bar/MyCustomPython" "$0" "$@"
    exit 127
    fi
    '''
    import …

It works like magic. All you have to do is set $VENV to evaluate to the
copy of python you need it to execute with.

Unfortunately, I think there _is_ a slight pause as a result. That is
totally acceptable for the portability it provides when someone needs to
use these scripts down the road.