Implement proper csrf on Flask_WTForms
I am not sure if I have it right, but I believe I do, and it _seems_ to
work, though I never though it didn't.

I'm a little bothered by this warning:

>./home/malan/projects/icc/icc/venv/lib/python3.7/site-packages/wtforms/form.py:212:
>FlaskWTFDeprecationWarning: "csrf_enabled" is deprecated and will be
>removed in 1.0. Set "meta.csrf" instead.
>`   return type.__call__(cls, *args, **kwargs)`

considering I have nothing to do with that code and wouldn't be able to
figure it out if I tried. May have to ask on StackOverflow what to do
about it.