Convert enum system to rely on an EnumMixin
This might still be broken.

Chiefly, 2 tests fail and I can't figure out why. They can't find a
value that they _should_ be able to find.

This commit is an initial commit on the path to better testing to fix
this problem (that works in the actual app, I don't understand why it
doesn't in the tests.