Initial model conversion
Essentially, I'm creating a new proxy/association object called
'LineAttribute' that will allow me to create an infinite list of line
attributes.

There's no reason to convert the emphasis status mechanism to the
attribute list model, or even move it to a new table, because I'm going
to continue to access it the way I have been.

Regarding the line attributes conversion, though, the association object
is interesting and not at all according to my standard mode by the
following features:

1. It has no id (it's really unnecessary, I believe, in this case.)
2. It uses association_proxies (which I have just learned about) to make
   the enum model more transparent. I'll never have to directly access
   the enum object itself, so no line.attr.enum.enum

It specifically has the following pieces of information that make this
work:

On the enum object:

1. The enum itself (which we use as a code in our CSS classes).
2. The display version of the enum (which we use to classify line
   locations/read views in the Table of Contents

On the assoc object:

1. Pointers to both the enum object and the line
2. A num integer column, so that if it is a `Chapter` attribute, I know
   the chapter number the line is contained in.
3. A precedence integer column, so that we can know the actual hierachy.
   The highest level starts at 1 and climbs from there. This allows us
   to define infinite levels of hierarchy without too much recursion
   work. Working this out on the Table of Contents might be tough.
3. A primary boolean column, so that I know if it is the header for the
   section. That is to say, for actual Chapter headers, or Book headers,
   etc, we flip it to true. All other attributes, this will be false.
4. An association_proxy to the enum.enum
5. An association_proxy to the enum.display

This is going to be far more complex a transition than I initially
suspected. The following steps will be involved:

1. I'm going to have to change the processor mechanism itself, which
   will be a good opportunity to refactor it into, perhaps, a more
   functional paradigm (at the very least I'd like to define more
   functions and call them in an `if __name__` block), cleaning it up.
2. I'm going to have to tweak the insertlines.py script to read the new
   processor output. This actually seems like it will be easier than
   (1).

At the end of all this, I'm going to clean up some other methods. I can
eliminate, for instance, my complex voter/votes relationships by just
defining association_proxies. I suspect similar things can be done in a
number of locations.

It pays to read the docs.