Refactor processor.py to new line attribute architecture
This new architecture is certainly more modulara. I prefer it, in many
ways. But I am not at *all* certain that it's easier to maintain.

That said, I'm not excited at all about rewriting my classic imperative
scripts, so perhaps this *is* the better method.

I've composed the problem into two master functions.

Each function has a class called Switch that it instantiates into an
object to process the file.

The `Switch` maintains the state and context for the given run,
(`readin`, and `readout`, are the two methods). It uses dictionaries of
lambdas composed in different ways to perform the basic analysis and
processing.

`readin`'s `Switch` has an incredibly awesome Turing Machine like
dictionary thing. Each key points to a value which is the successive
state of a class attribute. Every time I process the Turing Machine, it
just updates itself to the next value. It works fantastically.

I'm just a little worried about the conceptual complexity of the thing.
I also don't like how many complex dictionary comprehensions and lambdas
there are.

But at least it's black-boxified!

I still need to convert `insertlines.py` to be able to handle the new
modelling. I need to update `preprocessor.py` and the other insert
scripts to be more maintainable and functional. I need to, after I
update `insertlines.py`, go through the entire app and update tons of
stuff for processing chapters, etc., to use the new model.

Much still to do.

This is a massive refactoring. Bigger than I expected. And it's working
out.