Rewrite insertlines.py to handle the new line attributes architecture.
I have made considerable changes in this commit.

First, `insertlines.py` is now fairly functional. It's a lot cleaner. I
came up with the idea of, like with `Wiki` and `Annotation` creation,
making `LineAttribute` creation transparent with auto-creating
`LineEnums`, but it results in way too many db queries and slows the db
population down to a crawl. Since we only do this during the execution
of the `insertlines.py` script, there's no reason to create the
transparency issue. Just leave it in `inserlines.py`.

There were several bugs in the content models. They have been fixed in
this commit.

I have processed both wap and hod with the processor and added them to
`data/texts`.