:boom: Fix tests, mv inserts to inserts directory
I am modularizing the inserts, making them more functional, better
written. Unfortunately, the `conftest.py` can't use
`insertannotations.py` functions like I was hoping, because the
annotations file for the data isn't designed for it. It was designed to
represent user annotations while my `insertannotations.py` script was
written with author annotations in mind.

I believe there are broken changes in this commit. In fact, I know there
are. Because the `[re]populate.sh` and `recreate.sh` scripts are not
expecting the changes to the inserts.