Change annotationflags to aflags in templates
This was in the process of attempting to eliminate the need to populate
the annotationflags variable for every view that displays annotations.
It was a waste of time, and I was unsuccessful.

I might as well leave these as they are for now. I need to overhaul the
entire annotation flag system anyway.