Eliminate tag option & fix missing function in read
The elimination of the tag limitation option on the read view is purely
temporary. I want to implement a much more powerful version that can
intersect or union multiple tags. The query is simple enough in the
database, or at least the function I implemented, but the user interface
is going to be difficult and possibly require javascript to formulate
the URI. Probably not. The issue is mostly *eliminating* tags from a
current intersection/union, and not adding them. To add them, all you
have to do is add them. To remove them, I have no clue.