Offload user_flrs table to the relationship definition
This is cleaner, in my opinion. I only use it for one relationship
anyway, so there's no reason to have this thing defined as a separate
table.

I might end up doing this for all the tables.

Next, though, I'm going to see about defining a `FollowableMixin` to
avoid defining the relationships at all!