Create FollowableMixin and port everything to it
So the biggest obstacle to maintainability is anti-DRY code. Well, we
had that, in the follower system. I was obviously doing something wrong.

Now there's a followable mixin that automatically creates the table for
following and creates the relationship and backref any time it's
applied.

In the process of doing this I had to change the tablename naming schema
from the default (which I'm pretty sure is set by flask-migrate and I'm
good on that, I'll be moving to alembic when the whole project is over).

This means there are *probably* breaking changes in this commit. I'll
have to do some code review (I haven't even tested) so, well, no
problem.

The table naming scheme is now simply the class name lowercased.

Since there are only two tables in the table module right now I may
eventually eliminate that entire file. We'll see how it goes.

Next, we're going to work on that Votable mixin!