Create FlagMixin
This is nice I think I'm beginning to go overboard though.

The next logical step *would* be the `FlaggableMixin` that automatically
creates a `FlagMixin` inheriting class. But that would be taking things
a step too far in my opinion. Because ultimately I only really have two
flaggable objects for right now, though I should add flagging to Wikis
and both Requests. But furthermore, I need the `Annotation`-flagging
system to be capable of some nifty stuff down the road and composing a
`FlaggableMixin` would make that more difficult.

The main reason why I would want a `FlaggableMixin` is to add a
VotableMixin. It would be nice not to have to keep writing identical
`upvote()` and `downvote()` methods. But they're *not* all identical.
We'll see how I feel in a bit.