Replace outdated references (see body)
I changed every reference to `<cls>FlagEnum` to `<cls>Flag.enum_cls`. I
still have uses of `<cls>FlagEnum` but they are not so trivially
replacable (specifically, in my bootstrap insert scripts that rely on a
classes dictionary to add enums to the db). This will eventually be
mitigated, but for now it will stand.