Initial version of the votablemixin and repchange stuff
I created a VotableMixin with the upvote, downvote, and rollback. In
order to make these generic enough, I also added repchange and
rollback_repchange methods to User and renamed my first and default
repchange enums to correspond with the classname. I applied all this to
Annotation.

As I type this I've thought of a few better ideas. I already know I'm
going to have to make changes to make this thing work with edits, etc.

But this is a good first start.