ToC is now collapsed/expandable
I need a progressive enhancement on this. So still somewhat unsure how
to do that. Imagine it would be to not set them to display: none; in the
css but to do it manually on page load with javascript. This, however,
is more complicated than I initially thought, because i technically
don't have a universal script loader. I need to clean that up.