Finalize the annotate button
The annotate button is no longer an actual button. It's now just a
stylized div.

To take care of the timing issues and the various problems I've
encountered with line numbers and character numbers, etc. is the
following:

First, all of the events that detect when a user is highlighting text
activate a function, `show()`, that *simply* checks whether text is
highlighted and shows or disappears the annotate button.

Next, because those events are fairly irregular, we do not check the
numbers *until* the pointer enters the hitbox on the annotate button.
This creates/overwrites global variables for this.

Finally, *when* the button is clicked, the url is created and
location.href is changed.

This seems to be fairly robust and took a while to get to.

Next I have to try to eliminate the POST method on the default form and
change it to a GET method and in doing so change the url for the
annotate method to *not* include any line numbers, etc. Instead we will
be offloading that into the ? string.