Finish the ajax annotation voting
That turned out to be a doozy but I eventually got it.

Had to seriously modify the voting system. For one thing, I had no way
of knowing what had happened in the course of voting except for flashed
messages. So modifying the buttons and weight on the displayed
annotation was impossible without implementing something that would tell
me.

This meant I created an array of messages called situ that I use to pass
messages about what had happened.

This method is in fact stupid and overly complicated. I should just
return a tuple of success, status and just jsonify that instead.

TBD on next commit.

Anyway, this sucker works and I like it that way.