Selector in editor complete
So we're still having trouble in edge cases (namely the beginning and
ending of the book).

This doesn't seem to be that difficult but I'm struggling with a rather
strange bug.

There was a really nutso weird bug on annotate that I fixed in the route
(I was only getting the first line, lol). But that's gone.

So anyway, the biggest problem rightnow is the corner cases. The
beginning and the ending. And. yeah. I don't know what's going on with
that yet.