Make comments votable (see body)
Okay, so it turns out my legwork earlier was totally worth it. Making
something votable is totes easy. And making it reputation-changing is
also totes easy. Mixins are awesome.

Larger problems stem from the following:

I would like to algorithmically create voting routes (obvs this isn't
going to work with all of them, so you can override those? Or have a
dunder variable for whether to create it or not.

I also need to make this _vote.js work for all kinds of voting. So this
needs to become more generic.

I did make the _weight.html and _vote.html more generic. Those are
pretty good now. You have to set somevariables before calling them
(honestly that gets a little obnoxious and I would like to speed it up
or shorten it) but it's better than having a _vote.html for every
votable entity.

I deleted _weight.html from the requests templates so that is a breaking
change.

All in all very productive day, almost done with the comments. I really
think the _vote.js thing is the end of them. Well, then I'd like to do
the starrable, which isn't that... difficult? I guess I'll make a new
rule on comments that when they reach a certain vote threshhold they'll
be starred? I don't know. Nah, this is stupid. I'll leave it to votes
for now.

So just fix _vote.js