Reduce all voting routes to two for all
Now we have ajax voting and regular voting abstracted to two routes that
cover up and down votes for all VotableMixin objects.

I'm quite proud of this. It's not done yet because I still have to do
the ajax javascript to handle all the voting, but it works like a boss.

I love dynamic languages. Love love love them. Less LoC.

I can't really come up with any reason to describe a line by line change
except this:

There is one main.vote route for all voting that is ajax free. It highly
leverages the abstract nature of VotableMixin.

There is one ajax.vote route for all ajax voting that also highly
leverages the abstract nature of Votable Mixin.

All of the _weight.html includes have been reduced to a single include
in the generic templates/includes directory. All they require is that
entity be set to the entity.

All of the _vote.html includes have been reduced to a single include in
the generic templates/includes directory. All they require is that
prefix be set (just a single letter for the id) and ident be set (the
entity's id).

One thing I am tremendously proud of in this new voting system is how I
reduced duplicated code in the VotableMixin by using a central voteprep
method and *furthermore* that I learned to use !(a ^ b) as an IFF
method. I really liked that little bit. (if the vote is up and we are
upvoting, rollback; if the vote is down and we are downvoting,
rollback).

Anyway, proud of this change.