Fix expand selection
Okay, so little problem kicking down the road: right now the context is
not defined right.

This is going to be my first database migration after launch.

Going to have to add a field to edit that points to the toc.

The reason is this: I can't define context on an annotation so that it
excludes lines before or after a toc (when the annotation is at the
beginning or end of a toc) if I *don't* do this.

It will also simplify some code. For instance, right now I have a filter
on edit that prevents population of lines past the actual context of an
annotation (i.e., within the toc). But that's just a temporary fix